+++ checks/AZS/AZS.go
@@ -7,6 +7,7 @@ import (
 	AZS001 "github.com/katbyte/azproviderlint/checks/AZS/AZS001_typed_sdk_model_64bit_types"
 	AZS002 "github.com/katbyte/azproviderlint/checks/AZS/AZS002_schema_default_type_mismatch"
 	AZS003 "github.com/katbyte/azproviderlint/checks/AZS/AZS003_schema_allows_empty_block"
+	AZS004 "github.com/katbyte/azproviderlint/checks/AZS/AZS004_enum_validation_missing_values"
 )
 
 // Checks contains all AZS (schema & typed SDK model) analyzers.
@@ -14,4 +15,5 @@ var Checks = []*analysis.Analyzer{
 	AZS001.Analyzer,
 	AZS002.Analyzer,
 	AZS003.Analyzer,
+	AZS004.Analyzer,
 }